Added rate limiters and more Auth and cookie notice

+++ app/api/bugs/delete/route.js
@@ -1,14 +1,23 @@
 import {NextResponse} from 'next/server'
 import clientPromise from "@/lib/mongo/index";
 import Bug from "@/models/bugSchema"
+import { limiter } from "../../config/limiter";
 
 export async function POST(req,res)
 {
     try {
+        const limit = await limiter.removeTokens(1)
+        if(limit > 0)
+        {
         await clientPromise();
         const body = await req.json()
         await Bug.findByIdAndDelete(body);
         return NextResponse.json({message: "bug deleted", status : 201})
+        }
+        else
+        {
+            return new NextResponse({},{status: 429, statusText: "Too Many Requests"})
+        }
     } 
     catch (err) 
     {
